# Description

This adds the state size tracking mechanism to InMemorySorobanState. The
size includes the size of the parsed Wasm modules and thus has to be
recomputed on the protocol and relevant setting upgrades. We'll start
tracking the state size as soon as we're in Soroban protocol, but will
only use it in p23. Because of that we can use p23 host to compute the
memory cost for contract code entries eagerly.

This does a few non-trivial things on upgrades:

- When we perform a memory cost settings upgrade (even in p22) we
recompute the contract code in-memory size. If we are in p23+, we also
fill the state size window with the current in-memory state size.
- When we upgrade to p23, we immediately fill the state size window with
the current in-memory state size, but don't recompute the contract code
size (it's already up to date)
- When we upgrade to p24+, we recompute the contract code size and fill
the state window with the current in-memory state size

All these cases are covered in the respective new tests.

There is also a bunch of test-related TODOs here that can be addressed
later - they shouldn't influence the change coverage significantly.

# Checklist
- [ ] Reviewed the
[contributing](https://github.com/stellar/stellar-core/blob/master/CONTRIBUTING.md#submitting-changes)
document
- [ ] Rebased on top of master (no merge commits)
- [ ] Ran `clang-format` v8.0.0 (via `make format` or the Visual Studio
extension)
- [ ] Compiles
- [ ] Ran all tests
- [ ] If change impacts performance, include supporting evidence per the
[performance
document](https://github.com/stellar/stellar-core/blob/master/performance-eval/performance-eval.md)

+++ src/bucket/HotArchiveBucketIndex.h
@@ -94,7 +94,7 @@ class HotArchiveBucketIndex : public NonMovableOrCopyable
     uint32_t
     getPageSize() const
     {
-        return mDiskIndex.getPageSize();
+        return static_cast<uint32_t>(mDiskIndex.getPageSize());
     }
 
     IterT
